Payment Service

+++ src/main/java/com/restaurant/dto/PaymentDisplayDTO.java
@@ -6,7 +6,7 @@ import java.sql.Timestamp;
 
 @Data
 public class PaymentDisplayDTO {
-    private int clientId;
+    private Long clientId;
     private String clientName;
     private Timestamp time;
     private double cost;
