+++ src/main/java/com/restaurant/entity/Payment.java
@@ -4,9 +4,7 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import lombok.Builder;
-import lombok.Data;
-import lombok.ToString;
+import lombok.*;
 
 import java.sql.Timestamp;
 
@@ -18,7 +16,7 @@ public class Payment {
     @Column(name = "id")
     private Long id;
     @Column(name = "client_id")
-    private int clientId;
+    private Long clientId;
     @Column(name = "time")
     private Timestamp time;
     @Column(name = "cost")
